% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postino.R
\name{browse}
\alias{browse}
\title{Browse saved frames}
\usage{
browse(x, ..., delay = 1000, baseurl = NULL)
}
\arguments{
\item{x}{Postino instance.}

\item{...}{Files to browse.}

\item{delay}{Slide delay in milliseconds.}

\item{baseurl}{Base URL for the viewer. If \code{NULL}, the stored URL is used.}
}
\value{
The viewer URL is invisibly returned.
}
\description{
Launches the viewer in a new browser window to display saved frames as a slideshow.
}
\seealso{
Other Postino-methods: 
\code{\link{clear}()},
\code{\link{lst}()},
\code{\link{post}()},
\code{\link{shutdown}()}
}
\concept{Postino-methods}
