% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postino.R
\name{clear}
\alias{clear}
\title{Clear all saved frames}
\usage{
clear(x, ...)
}
\arguments{
\item{x}{Postino instance.}

\item{...}{Not used.}
}
\description{
Deletes all saved JSON files.
}
\seealso{
Other Postino-methods: 
\code{\link{browse}()},
\code{\link{lst}()},
\code{\link{post}()},
\code{\link{shutdown}()}
}
\concept{Postino-methods}
