% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{etegami-package}
\alias{etegami}
\alias{etegami-package}
\title{etegami: HTML5 Viewer for nativeRaster}
\description{
An experimental viewer for 'nativeRaster' images in R, rendered on an 'HTML5' canvas element via browsers.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
