% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postino.R
\name{lst}
\alias{lst}
\title{List saved frames}
\usage{
lst(x, ...)
}
\arguments{
\item{x}{Postino instance.}

\item{...}{Not used.}
}
\value{
A character vector of file names.
}
\description{
Lists saved JSON files.
}
\seealso{
Other Postino-methods: 
\code{\link{browse}()},
\code{\link{clear}()},
\code{\link{post}()},
\code{\link{shutdown}()}
}
\concept{Postino-methods}
