% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mask_plot}
\alias{mask_plot}
\title{Wrap the base plot function to auto-capture frames}
\usage{
mask_plot(postino)
}
\arguments{
\item{postino}{A \link{Postino} object used to capture and save frames.}
}
\value{
A function compatible with \code{plot()} that also triggers capture.
}
\description{
Returns a plotting function that behaves like \code{\link[base:plot]{base::plot()}}
but also captures the result using a \link{Postino} instance.
}
