% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postino.R
\name{post}
\alias{post}
\title{Capture and save a frame}
\usage{
post(x, ...)
}
\arguments{
\item{x}{Postino instance.}

\item{...}{Not used.}
}
\value{
The name of the saved JSON file.
}
\description{
Captures a raster image using the associated graphics device and saves it as a JSON file.
}
\seealso{
Other Postino-methods: 
\code{\link{browse}()},
\code{\link{clear}()},
\code{\link{lst}()},
\code{\link{shutdown}()}
}
\concept{Postino-methods}
