% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postino.R
\name{Postino}
\alias{Postino}
\title{An S7 class for managing raster capture and viewer interaction}
\usage{
Postino(
  serve_dir = character(0),
  capture_device = function() NULL,
  id_counter = function() NULL,
  mode = character(0),
  server_config = list()
)
}
\arguments{
\item{serve_dir}{Directory where JSON files are saved.}

\item{capture_device}{A graphics device function to use for capturing frames. See \code{\link[ragg:agg_capture]{ragg::agg_capture()}}.}

\item{id_counter}{A function to generate unique IDs for frames.}

\item{mode}{\code{"batch"} (default) or \code{"live"}.}

\item{server_config}{A list of configuration settings out of \code{\link[servr:httd]{servr::httd()}}.}
}
\description{
\code{Postino} objects manage the capture of \code{nativeRaster} images, saving them as JSON,
and controlling a browser-based viewer interface. For usage, see \code{\link[=setup]{setup()}}.
}
\examples{
\dontrun{
# Instead of using `setup()`, you can create a Postino instance manually:
postino <- Postino(
  serve_dir = "Postino-dir",
  capture_device = ragg::agg_capture(),
  id_counter = counter(),
  mode = "batch",
  server_config = servr::httd("Postino-dir", browser = FALSE)
)
}
}
