% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrjson-io.R
\name{read_nrjson}
\alias{read_nrjson}
\title{Read a nativeRaster JSON file}
\usage{
read_nrjson(path)
}
\arguments{
\item{path}{Path to the JSON file.}
}
\value{
A \code{nativeRaster} object with \code{id} attribute.
}
\description{
Loads a gzipped JSON file that encodes a nativeRaster object and restores it
as an R \code{nativeRaster}.
}
