% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup}
\alias{setup}
\title{Initialize a Postino instance and launch viewer server}
\usage{
setup(
  out_dir = tempdir(),
  capture_device = ragg::agg_capture(),
  id_counter = counter(),
  mode = c("batch", "live"),
  ...
)
}
\arguments{
\item{out_dir}{Output directory. Defaults to a temporary directory.}

\item{capture_device}{A function that returns a \code{nativeRaster} object.}

\item{id_counter}{A function to generate frame IDs.
Defaults to \code{\link[=time_stamp]{time_stamp()}}.}

\item{mode}{Viewer mode. Either \code{"batch"} or \code{"live"}.
If \code{"live"} mode, the viewer will automatically launch
every time \code{post()} is called.}

\item{...}{Additional arguments passed to \code{\link[servr:httd]{servr::httd()}}.}
}
\value{
A \code{Postino} object.
}
\description{
Sets up a temporary directory, copies viewer assets, starts a local server,
and returns a \code{Postino} instance for managing raster output and display.
}
\note{
Note that calling \code{setup()} open a new graphics device via \code{\link[ragg:agg_capture]{ragg::agg_capture()}}
and launch a local HTTP server in the background using \code{\link[servr:httd]{servr::httd()}}.
}
