% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postino.R
\name{shutdown}
\alias{shutdown}
\title{Stop the httd server}
\usage{
shutdown(x, ...)
}
\arguments{
\item{x}{Postino instance.}

\item{...}{Not used.}
}
\description{
Stops the viewer's HTTP server, if it is running.
And also, this always calls \code{\link[grDevices:dev]{grDevices::dev.off()}} internally.
}
\seealso{
Other Postino-methods: 
\code{\link{browse}()},
\code{\link{clear}()},
\code{\link{lst}()},
\code{\link{post}()}
}
\concept{Postino-methods}
