% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_stamp}
\alias{time_stamp}
\title{Generate a timestamp string}
\usage{
time_stamp()
}
\value{
A character string representing the timestamp.
}
\description{
Returns the current system time in \code{"\%Y\%m\%d_\%H\%M\%S"} format.
}
\keyword{internal}
