% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrjson-io.R
\name{write_nrjson}
\alias{write_nrjson}
\title{Write a nativeRaster object to JSON}
\usage{
write_nrjson(x, id, path = paste0(id, ".json"))
}
\arguments{
\item{x}{A \code{nativeRaster} object.}

\item{id}{A string identifier for the image.}

\item{path}{File path to write to. Defaults to \code{paste0(id, ".json")}.}
}
\value{
The \code{path} is invisibly returned.
}
\description{
Encodes a \code{nativeRaster} object as gzipped JSON
and saves it to disk. The file can be viewed in a compatible browser viewer.
}
