% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup}
\alias{setup}
\title{Initialize a Postino instance and launch viewer server}
\usage{
setup(
  out_dir = tempdir(),
  capture_device = ragg::agg_capture(width = 720, height = 576),
  id_counter = counter(),
  mode = c("batch", "live"),
  ...
)
}
\arguments{
\item{out_dir}{Output directory where viewer assets are copied into.
Defaults to a temporary directory.}

\item{capture_device}{A function that returns a \code{nativeRaster} object.}

\item{id_counter}{A function to generate frame IDs.}

\item{mode}{Viewer mode. Either \code{"batch"} or \code{"live"}.
If \code{"live"} mode, the viewer will automatically launch
every time \code{post()} is called.}

\item{...}{Additional arguments passed to \code{\link[servr:httd]{servr::httd()}}.}
}
\value{
A \code{Postino} object.
}
\description{
Sets up a new subdirectory including viewer assets, starts a local server,
and returns a \code{Postino} instance for managing raster output and display.
\code{setup()} initializes a new graphics device via \code{\link[ragg:agg_capture]{ragg::agg_capture()}}
and starts a local HTTP server in the background using \code{\link[servr:httd]{servr::httd()}}.
}
